+++ bitrise.yml
@@ -3,9 +3,11 @@ default_step_lib_source: https://github.com/bitrise-io/bitrise-steplib.git
 
 app:
   envs:
-  - BITRISE_STEP_ID: expo-detach
-  - BITRISE_STEP_VERSION: "0.9.0"
   - BITRISE_STEP_GIT_CLONE_URL: https://github.com/bitrise-steplib/steps-expo-detach.git
+
+  # Define it in .bitrise.secrets.yml
+  - USER_NAME: $USER_NAME
+  - PASSWORD: $PASSWORD
 
 workflows:
   test:
